Add files via upload
International Mother's Day

Aperson who deserves to be the most cherished, loved, respected, admired, cared for, and every other thing that can be showered upon someone, is the only one in the whole world. One of the most admirable and strong bonds that one can have in today’s time is with their mothers. That is a relationship incomparable. A mother’s love is immeasurable.

As said, the highest amount of love and respect, if it should be given to someone, it should be to mothers. Not just your mother but to every single woman who is, was, and is going to be a mother. The woman who gave birth to you nourished you to be able to open your eyes when you enter the world with her blood, gave you the strength, and stood as your backbone just so you could stand out there and shine, she deserves to be given the priority in everything.


 
No matter how many days, months or years it takes, repaying our mothers for what they have done and can do for us is just not possible. To try and do that just a bit, there’s a whole day dedicated to mothers; Mother’s Day.


Mother’s Day is celebrated on the second Sunday of May, every year. This day isn’t just dedicated to the mothers but motherhood and maternal bonds are also a part of the reason for the day being celebrated. The day is celebrated worldwide. However, it was in the US in 1907 that it was first introduced and celebrated. A woman named Anna Jarvis was the reason that the world celebrated Mother’s day today. It was on the day of her mother's funeral in 1905, that she decided to establish the campaign for celebrating a day for mothers every year. It was then that she organized the first Mother’s Day service of worship, after two years. The church in which this was organized, Andrews Methodist Episcopal Church in West Virginia, now serves as the international Mother’s Day Shrine. Anna Jarvis was majorly focused on injured officers of the American Civil War, regardless of whose side were they on. She was a peace activist. To address general medical problems, she introduced Mother’s Day Work Clubs. She, along with another dissident, Julia Ward Howe had been encouraging the making of a "Mother's Day For Peace". It was the day that was going to be when moms would ask that their children and husbands weren’t under the compulsion to go to war and not be killed. Anna Jarvis believed that ‘the individual who has accomplished more for you than anybody on the planet’, and for her, it was the mothers of the world. Celebrating a day for mothers solely and tributing everything to the person who will love you the most ever in your entire life is one of the greatest things you can do for that person, for your mother. However, Anna Jarvis’ initiation of establishing a Mother’s Day in the world didn’t have such positive and appreciative thoughts when the campaign was introduced. In 1908, the government of the US, the Congress, rejected the idea of initiating the campaign, taking it further as a joke and claiming that if they agreed to set in a day like that, more like a day for mothers-in-law will also come up in line. This surely didn’t amuse Anna at all as after three years of utmost dedication and hard work to finish what she started, the US government established Mother’s Day to be on the second Sunday of May every year in 1911.